Updated FF usage
