+++ sdk/spring/spring-cloud-azure-appconfiguration-config/src/main/java/com/azure/spring/cloud/appconfiguration/config/implementation/AppConfigurationSnapshotPropertySource.java
@@ -2,9 +2,13 @@
 // Licensed under the MIT License.
 package com.azure.spring.cloud.appconfiguration.config.implementation;
 
+import static com.azure.spring.cloud.appconfiguration.config.implementation.AppConfigurationConstants.FEATURE_FLAG_PREFIX;
+import static com.azure.spring.cloud.appconfiguration.config.implementation.AppConfigurationConstants.FEATURE_MANAGEMENT_KEY;
+
 import java.util.List;
 
 import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
+import com.azure.spring.cloud.appconfiguration.config.implementation.http.policy.TracingInfo;
 import com.fasterxml.jackson.core.JsonProcessingException;
 
 /**
@@ -44,5 +48,17 @@ final class AppConfigurationSnapshotPropertySource extends AppConfigurationAppli
         throws JsonProcessingException {
         // Feature Flags are only part of this if they come from a snapshot
         processFeatureFlag(key, setting, trimStrings);
+    }
+    
+    protected void processFeatureFlag(String key, FeatureFlagConfigurationSetting setting, List<String> trimStrings) {
+        TracingInfo tracing = replicaClient.getTracingInfo();
+        featureConfigurationSettings.add(setting);
+        FeatureFlagConfigurationSetting featureFlag = setting;
+
+        String configName = FEATURE_MANAGEMENT_KEY + setting.getKey().trim().substring(FEATURE_FLAG_PREFIX.length());
+
+        AppConfigurationFeatureManagementPropertySource.updateTelemetry(featureFlag, tracing);
+
+        properties.put(configName, AppConfigurationFeatureManagementPropertySource.createFeature(featureFlag));
     }
 }
